deadpool_redis

Struct Config

pub struct Config {
    pub url: Option<String>,
    pub connection: Option<ConnectionInfo>,
    pub pool: Option<PoolConfig>,
}

Configuration object.

Example (from environment)

By enabling the serde feature you can read the configuration using the config crate as following:

REDIS__URL=redis.example.com
REDIS__POOL__MAX_SIZE=16
REDIS__POOL__TIMEOUTS__WAIT__SECS=2
REDIS__POOL__TIMEOUTS__WAIT__NANOS=0

#[derive(serde::Deserialize)]
struct Config {
    redis: deadpool_redis::Config,
}

impl Config {
    pub fn from_env() -> Result<Self, config::ConfigError> {
        let mut cfg = config::Config::builder()
           .add_source(config::Environment::default().separator("__"))
           .build()?;
           cfg.try_deserialize()
    }
}

Fields
url: Option<String>

Redis URL.

See Connection Parameters.

connection: Option<ConnectionInfo>

[redis::ConnectionInfo] structure.

pool: Option<PoolConfig>

Pool configuration.

Implementations

impl Config

pub fn create_pool( &self, runtime: Option<Runtime>, ) -> Result<Pool, CreatePoolError>

Creates a new Pool using this Config.

Errors

See CreatePoolError for details.

pub fn builder(&self) -> Result<PoolBuilder, ConfigError>

Creates a new PoolBuilder using this Config.

Errors

See ConfigError for details.

pub fn get_pool_config(&self) -> PoolConfig

Returns deadpool::managed::PoolConfig which can be used to construct a [deadpool::managed::Pool] instance.

pub fn from_url<T: Into<String>>(url: T) -> Config

Creates a new Config from the given Redis URL (like redis://127.0.0.1).

pub fn from_connection_info<T: Into<ConnectionInfo>>( connection_info: T, ) -> Config

Creates a new Config from the given Redis ConnectionInfo structure.